Updated the testexposure model + wrote testcase for running schema mapping to suppasri
